flitify-webpanel/backend/requirements.txt
requests
flask
waitress
pymongo
pyjwt
bcrypt
